requirements.txt
fastapi
openpyxl
pandas
python-dotenv
requests
requests-toolbelt
requests_toolbelt
uvicorn[standard]
